install pocketbase + update frontend (add themes and footer)
