+++ frontend/src/Components/Navbar/Navbar.js
@@ -20,37 +20,14 @@ const Navbar = () => {
     <a class="btn btn-ghost normal-case text-xl" href="#" data-uri="/" >BattleShip</a>
     </div>
     <div class="navbar-end">
+
       <button class="btn btn-ghost" href="#" data-uri="/game">Game</button>
       <button class="btn btn-ghost" href="#" data-uri="/new">New Page</button>
       <button class="btn btn-ghost" href="#" data-uri="/login">Login</button>
       <button class="btn btn-ghost" href="#" data-uri="/register">Register</button>
       
 
-      <div class="dropdown dropdown-hover">
-        <label data-theme="luxury" tabindex="0" class="btn m-1">Themes</label>
-          <ul tabindex="0" class="dropdown-content z-[1] menu p-2 shadow bg-base-100 rounded-box w-52">
-
-            <li><a data-theme="luxury">luxury</a></li> 
-            <li><a data-theme="forest">forest</a></li>
-            <li><a data-theme="dark">dark</a></li> 
-
-            <li><a data-theme="synthwave">synthwave</a></li>
-            <li><a data-theme="halloween">halloween</a></li>
-            <li><a data-theme="black">black</a></li>
-
-            <li><a data-theme="dracula">dracula</a></li>
-            <li><a data-theme="business">business</a></li>
-            <li><a data-theme="night">night</a></li>
-
-            <li><a data-theme="coffee">coffee</a></li>
-            <li><a data-theme="aqua">aqua</a></li>
-            <li><a data-theme="garden">garden</a></li>
-
-            <li><a data-theme="retro">retro</a></li>
-            <li><a data-theme="valentine">valentine</a></li>
-            <li><a data-theme="wireframe">wireframe</a></li>
-          </ul>
-      </div>
+      
 
     </div>
   </div>
@@ -58,31 +35,7 @@ const Navbar = () => {
 
   navbarWrapper.innerHTML = navbar;
 
-  // ----------------------------------------------changement de theme----------------------------------------------------------------------
-
-  // Sélectionnez tous les liens du dropdown
-  const dropdownLinks = document.querySelectorAll('.menu li a');
-
-  // Définissez une fonction pour changer le thème
-  function changeTheme(theme) {
-    // Récupérez la balise <html>
-    const htmlTag = document.getElementById('htmlTag');
-
-    // Définissez la valeur de l'attribut data-theme de la balise <html>
-    htmlTag.setAttribute('data-theme', theme);
-
-  };
-
-  // Ajoutez un gestionnaire d'événements pour le clic sur les liens du dropdown
-  dropdownLinks.forEach(link => {
-    link.addEventListener('click', () => {
-      // Récupérez la valeur de l'attribut data-theme du lien
-      const themeValue = link.getAttribute('data-theme');
-
-      // Appelez la fonction pour changer le thème
-      changeTheme(themeValue);
-    });
-  });
+  
 
 };
