+++ frontend/src/index.html
@@ -11,6 +11,7 @@
   <body>
     <header><div id="navbarWrapper"></div></header>
     <main class="grid"></main>
-    <footer></footer>
+    <footer><div id="footerWrapper"></div></footer>
   </body>
+  
 </html>
