+++ frontend/src/index.js
@@ -1,8 +1,13 @@
 import './stylesheets/output.css';
 
 import Navbar from './Components/Navbar/Navbar';
+
+import Footer from './Components/Footer/Footer';
+
 import Router from './Components/Router/Router';
 
 Navbar();
+
+Footer();
 
 Router();
