+++ frontend/src/stylesheets/main.css
@@ -8,3 +8,11 @@ html {
     background-size: auto, auto;
     background-repeat: no-repeat, no-repeat;
 }
+
+
+  
+  footer {
+    position: fixed;
+    bottom: 0;
+    width: 100%;
+  }
